+++ README.md
@@ -47,9 +47,9 @@ generated by the compiler. While it is highly recommended, Typescript is NOT req
 houdini is available on npm:
 
 ```sh
-yarn add -D houdini houdini-preprocess houdini-compiler graphql
+yarn add -D houdini houdini-preprocess graphql
 # or
-npm install --save-dev houdini houdini-preprocess houdini-compiler graphql
+npm install --save-dev houdini houdini-preprocess graphql
 ```
 
 ## 🔧&nbsp;&nbsp;Configuring Your Application
@@ -59,7 +59,7 @@ If you don't already have an existing app, visit [this link](https://sapper.svel
 for help setting one up. Once you have a project and want to add houdini, execute the following command:
 
 ```sh
-npx houdini-compiler init
+npx houdini init
 ```
 
 This will create a few necessary files, as well as pull down a json representation of
@@ -98,7 +98,7 @@ to creating types for your documents. Running the compiler can be done with npx 
 in `package.json` and needs to be run every time a GraphQL document in your source code changes:
 
 ```sh
-npx houdini-compiler generate
+npx houdini generate
 ```
 
 ## 🚀&nbsp;&nbsp;Fetching Data
