--- packages/houdini-compiler/tsconfig.json
@@ -1,8 +1,0 @@
-{
-	"extends": "../../tsconfig",
-	"compilerOptions": {
-		"outDir": "build"
-	},
-	"include": ["src/**/*"],
-	"exclude": ["node_modules", "**/*.test.ts"]
-}
