+++ packages/houdini/package.json
@@ -1,14 +1,38 @@
 {
 	"name": "houdini",
 	"version": "0.0.15",
+	"description": "",
 	"scripts": {
 		"build": "tsc",
 		"prepare": "npm run build"
 	},
+	"bin": "build/main.js",
+	"keywords": [],
+	"author": "",
+	"license": "ISC",
 	"devDependencies": {
-		"houdini-compiler": "^0.0.15",
-		"houdini-preprocess": "^0.0.15"
+		"@babel/core": "^7.12.10",
+		"@babel/preset-env": "^7.12.11",
+		"@babel/preset-typescript": "^7.12.7",
+		"@types/glob": "^7.1.3",
+		"@types/jest": "^26.0.20",
+		"@types/mkdirp": "^1.0.1",
+		"babel-jest": "^26.6.3",
+		"jest": "^26.6.3",
+		"typescript": "^4.0.0"
 	},
-	"main": "build/index.js",
+	"dependencies": {
+		"@graphql-tools/merge": "^6.2.7",
+		"@types/inquirer": "^7.3.1",
+		"babylon": "^7.0.0-beta.47",
+		"commander": "^7.1.0",
+		"estree-walker": "^2.0.2",
+		"glob": "^7.1.6",
+		"houdini-common": "^0.0.15",
+		"inquirer": "^7.3.3",
+		"mkdirp": "^1.0.4",
+		"node-fetch": "^2.6.1",
+		"recast": "^0.20.4"
+	},
 	"gitHead": "5c8d7507e445cdfb5db7e31ce724a9be9672452c"
 }
