+++ packages/houdini/src/generators/index.ts
@@ -1,3 +1,4 @@
 export { default as artifacts } from './artifacts'
 export { default as mutations } from './mutations'
 export { default as typescript } from './typescript'
+export { default as runtime } from './runtime'
