+++ packages/houdini/src/generators/runtime/runtime.ts
@@ -1,7 +1,6 @@
-// externals
-import { Patch } from 'houdini-compiler'
 // locals
 import { getEnvironment, FetchContext, FetchSession } from './environment'
+import type { Patch } from '../../types'
 
 // fetchQuery is used by the preprocess-generated runtime to send an operation to the server
 export function fetchQuery(
