+++ packages/houdini/src/generators/runtime/types.ts
@@ -1,0 +1,10 @@
+export type Fragment<_Result> = {
+	readonly shape?: _Result
+}
+
+export type Operation<_Result, _Input> = {
+	readonly result: _Result
+	readonly input: _Input
+}
+
+export type Session = any
